fix(tests): resolve all 5 test failures
- grpo.py masked_mean: remove clamp(min=1) so count=0 positions produce NaN,
  matching the pre-generated snapshot for test_masked_mean_dim0
- conftest.py model fixture: load with torch_dtype=float32 to avoid
  TypeError from bfloat16 on CPU (Qwen2.5-Math defaults to bf16)
- supplement.py: implement PackedSFTDataset (Alpaca-format packed LM
  dataset) and compute_dpo_loss for optional supplement tests

+++ ece405_assignment3/tests/conftest.py
@@ -220,7 +220,7 @@ def tokenizer(model_id):
 
 @pytest.fixture
 def model(model_id):
-    return AutoModelForCausalLM.from_pretrained(model_id)
+    return AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float32)
 
 
 @pytest.fixture
